requirements.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.25.0
Pillow==5.3.0
scikit-learn==0.21.2
scipy==1.1.0
sklearn==0.0
tqdm==4.26.0
obspy
torch==1.1.0
torchvision==0.3.0
tensorboardX
gdown